Stp-M4: Zero-question walkthrough UX
The walkthrough is now an auto-fix-first narrated pass: 'Step 2 of 4 -
the database that stores your monsters' with a one-line 'already good'
when nothing needs doing, no per-component consent questions, no
between-step pauses, and failures collected onto the finish screen
instead of stopping the pass. The finish screen ends on the one action
only the player can do: the gear icon and the DeepSeek key, with the
cost ballpark stated plainly.

setup/user_messages.py (grandfathered at 1092 lines) becomes the
setup/messages/ package - default_path vs local-extras modules, all
under the ceiling, content moved verbatim - and its grandfather entry
is the first ever retired.
