+++ setup/messages/default_path.py
@@ -1,0 +1,260 @@
+"""
+Default-path user messages - everything a NEW PLAYER can ever be shown.
+Written for someone who has never used a terminal: short cards, the few
+choices that matter, no jargon. The local-LLM escape hatch messages
+live in their own modules and only surface via --local-extras.
+"""
+
+DEFAULT_PATH_MESSAGES = {
+    'nodejs_installation': [
+        "📋 NODE.JS NEEDS TO BE INSTALLED:",
+        "",
+        "Node.js runs the game's interface. Two ways to get it:",
+        "",
+        "EASIEST: close this window and double-click start_game.bat -",
+        "it can install Node.js for you automatically.",
+        "",
+        "By hand instead:",
+        "   1. Go to https://nodejs.org/",
+        "   2. Download the LTS version and run the installer",
+        "   3. The standard 'Next, Next, Finish' answers are all correct",
+        "   4. When it finishes, double-click start_game.bat again",
+        "     (freshly installed software isn't visible to windows that",
+        "      were already open, so the restart matters)",
+        "",
+    ],
+    'mysql_installation': [
+        "📋 INSTALLING MYSQL - WHAT TO DO:",
+        "",
+        "The MySQL download page just opened in your browser.",
+        "(If it didn't, go to: https://dev.mysql.com/downloads/mysql/)",
+        "",
+        "1. Download the INSTALLER for Windows (the MSI download, not the ZIP)",
+        "   and run the file it gives you.",
+        "",
+        "2. The wizard asks a lot of questions. Only these answers matter:",
+        "",
+        "     Setup type ...... 'Server only' is plenty",
+        "     Port ............ keep 3306",
+        "     Root password ... MAKE ONE UP and keep it handy -",
+        "                       this setup asks for it in a minute",
+        "     Windows Service . yes, and start at system startup",
+        "",
+        "   For everything else, accept what the wizard suggests.",
+        "",
+        "3. Let it finish (5-10 minutes). If a separate 'Configurator'",
+        "   window appears near the end, finish that too - it's the part",
+        "   that actually starts the server.",
+        "",
+        "This setup will wait here and check as soon as you're done.",
+        "",
+    ],
+    'mysql_service_start': [
+        "📋 STARTING MYSQL BY HAND:",
+        "",
+        "1. Press the Windows key + R together, type  services.msc",
+        "   and press Enter - a window called 'Services' opens.",
+        "2. Scroll to the entries starting with 'MySQL' (like MySQL84).",
+        "3. Right-click it and choose 'Start'.",
+        "4. So it starts by itself from now on: right-click it again,",
+        "   choose 'Properties', set 'Startup type' to 'Automatic', OK.",
+        "",
+        "If Windows says it can't start the service, a reinstall with the",
+        "wizard's suggested answers is usually faster than debugging it.",
+    ],
+    'mysql_troubleshooting': [
+        "📋 MYSQL TROUBLESHOOTING:",
+        "",
+        "Common issues and solutions:",
+        "",
+        "Issue 1 - Port 3306 in use:",
+        "1. Open Command Prompt as Administrator",
+        "2. Run: netstat -ano | findstr :3306",
+        "3. If another process is using port 3306:",
+        "   - Stop that process, or",
+        "   - Reconfigure MySQL to use a different port",
+        "",
+        "Issue 2 - MySQL won't start:",
+        "1. Check Windows Event Viewer:",
+        "   - Press Win+R → eventvwr.msc",
+        "   - Go to Windows Logs → Application",
+        "   - Look for MySQL errors",
+        "2. Common fixes:",
+        "   - Delete ib_logfile0 and ib_logfile1 from MySQL data directory",
+        "   - Run MySQL as Administrator",
+        "   - Check disk space (MySQL needs space for logs)",
+        "",
+        "Issue 3 - Permission errors:",
+        "1. Run Command Prompt as Administrator",
+        "2. Navigate to MySQL bin directory",
+        "3. Try: mysqld --install",
+        "4. Then: net start mysql",
+        "",
+        "Issue 4 - Configuration problems:",
+        "1. Locate my.ini file (usually in MySQL installation folder)",
+        "2. Check for syntax errors",
+        "3. Reset to default configuration if needed",
+        "",
+        "Last resort:",
+        "- Completely uninstall MySQL",
+        "- Delete all MySQL folders",
+        "- Reinstall with default settings",
+    ],
+    'database_password_setup': [
+        "📋 YOUR MYSQL PASSWORD:",
+        "",
+        "The game needs the MySQL password to store your monsters and",
+        "progress. It's the one you made up during the MySQL install",
+        "(the 'root password').",
+        "",
+        "You'll only be asked once - it's saved so the game can use it.",
+        "",
+    ],
+    'database_connection_troubleshooting': [
+        "📋 START MYSQL SERVER:",
+        "",
+        "Your MySQL server needs to be running for the game to work.",
+        "Here's how to start it:",
+        "",
+        "Method 1 - Windows Services (Recommended):",
+        "1. Press Win+R, type 'services.msc', press Enter",
+        "2. Look for 'MySQL80' or 'MySQL84' or 'MySQL'",
+        "3. If status shows 'Stopped', right-click → 'Start'",
+        "4. Status should change to 'Running'",
+        "",
+        "Method 2 - Command Line:",
+        "1. Open Command Prompt as Administrator",
+        "2. Try: net start MySQL80",
+        "3. Or try: net start MySQL84",
+        "4. Or try: net start MySQL",
+        "5. You should see: 'The MySQL service was started successfully'",
+        "",
+        "Method 3 - MySQL Workbench:",
+        "1. Open MySQL Workbench (if installed)",
+        "2. Look for 'Server Status' or 'Server Administration'",
+        "3. Click 'Start Server'",
+        "",
+        "💡 MySQL server must be running every time you play the game",
+    ],
+    'database_troubleshooting': [
+        "📋 GENERAL DATABASE TROUBLESHOOTING:",
+        "",
+        "Connection issues can have several causes:",
+        "",
+        "Issue 1 - Firewall blocking connection:",
+        "1. Check Windows Firewall settings",
+        "2. Allow MySQL through firewall (port 3306)",
+        "3. Or temporarily disable firewall to test",
+        "",
+        "Issue 2 - MySQL configuration problems:",
+        "1. Find my.ini file (usually in MySQL installation folder)",
+        "2. Check bind-address setting (should be 0.0.0.0 or commented)",
+        "3. Restart MySQL service after changes",
+        "",
+        "Issue 3 - User permissions:",
+        "1. Connect to MySQL as root",
+        "2. Check user permissions: SHOW GRANTS FOR 'root'@'localhost';",
+        "3. Grant permissions if needed: GRANT ALL PRIVILEGES ON *.* TO 'root'@'localhost';",
+        "",
+        "Issue 4 - Wrong host/port in .env:",
+        "1. Check .env file has correct settings:",
+        "   DB_HOST=localhost",
+        "   DB_PORT=3306",
+        "2. Verify MySQL is actually running on these",
+        "",
+        "Issue 5 - MySQL version compatibility:",
+        "1. Check MySQL version: mysql --version",
+        "2. Older versions may have different authentication",
+        "3. Try: ALTER USER 'root'@'localhost' IDENTIFIED WITH mysql_native_password BY 'your_password';",
+        "",
+        "Last resort - Reinstall MySQL:",
+        "1. Uninstall MySQL completely",
+        "2. Delete all MySQL folders and data",
+        "3. Reinstall with default settings",
+    ],
+    'database_manual_creation': [
+        "📋 MANUAL DATABASE CREATION:",
+        "",
+        "   Create the game database manually using one of these methods:",
+        "",
+        "   Method 1 - MySQL Workbench (Easiest):",
+        "      1. Open MySQL Workbench",
+        "      2. Connect to your local MySQL server",
+        "      3. In the query tab, run:",
+        "         CREATE DATABASE IF NOT EXISTS monster_hunter_game;",
+        "      4. Click the lightning bolt to execute",
+        "      5. Verify: You should see 'monster_hunter_game' in the schemas panel",
+        "",
+        "   Method 2 - Command Line:",
+        "      . Open Command Prompt",
+        "      2. Connect to MySQL:",
+        "         mysql -u root -p",
+        "      3. Enter your MySQL root password",
+        "      4. Create database:",
+        "         CREATE DATABASE IF NOT EXISTS monster_hunter_game;",
+        "      5. Verify:",
+        "         SHOW DATABASES;",
+        "         (You should see 'monster_hunter_game' in the list)",
+        "      6. Exit: EXIT;",
+        "",
+        "   Method 3 - phpMyAdmin (if you have XAMPP/WAMP):",
+        "      1. Open phpMyAdmin in browser",
+        "      2. Click 'Databases' tab",
+        "      3. Enter 'monster_hunter_game' as database name",
+        "      4. Click 'Create'",
+        "",
+        "   To test the database was created:",
+        "      1. Run this setup again",
+        "      2. Or manually test: mysql -u root -p -e \"USE monster_hunter_game; SELECT 1;\"",
+        "",
+        "💡 Database name must be exactly: monster_hunter_game",
+    ],
+    'basic_backend_diagnostic': [
+        "",
+        "🚨 CRITICAL: BASIC BACKEND SETUP FAILED 🚨",
+        "",
+        "⚠️  THE GAME CANNOT BE SETUP WITH THESE ISSUES ⚠️",
+        "",
+        "   Basic backend components are REQUIRED PREREQUISITES for all other setup steps.",
+        "   If you continue with these issues, ALL other components will fail:",
+        "",
+        "   Common problems and solutions:",
+        "",
+        "   Python Version Issues:",
+        "      • Python 3.8+ required - upgrade from https://python.org/downloads",
+        "      • Make sure to check 'Add Python to PATH' during installation",
+        "      • Restart command prompt after Python installation",
+        "",
+        "   pip Package Manager Issues:",
+        "      • pip should come with Python - if missing, Python installation is broken",
+        "      • Try: python -m ensurepip --upgrade",
+        "      • Or reinstall Python with default settings",
+        "",
+        "   Network Access Issues:",
+        "      • Check internet connection for downloading packages",
+        "",
+        "   Virtual Environment Issues:",
+        "      • Usually auto-created, failure indicates permission or disk space problems",
+        "      • Check you have write permissions in current directory",
+        "      • Antivirus may be blocking file creation",
+        "",
+        "   Environment File Issues:",
+        "      • .env.example template file may be missing from project",
+        "",
+        "   Why continuing is NOT recommended:",
+        "      • Other components depend on working Python environment",
+        "      • Installation will likely fail at later stages",
+        "      • May cause confusing error messages",
+        "      • Game won't work without proper backend setup",
+        "",
+        "   RECOMMENDED ACTION:",
+        "      1. Exit this setup now",
+        "      2. Fix the issues shown above",
+        "      3. Restart your command prompt",
+        "      4. Run this setup again",
+        "      5. The auto-setup should work once environment is fixed",
+        "",
+        "⚠️  ONLY CONTINUE IF YOU ARE AN EXPERIENCED DEVELOPER ⚠️",
+        "",
+    ],
+}
