+++ setup/messages/local_extras_gpu.py
@@ -1,0 +1,327 @@
+"""
+Local-extras messages: NVIDIA GPU and CUDA toolkit.
+Only shown by the --local-extras walkthrough (the unsupported local-LLM
+escape hatch) - a new player on the API-first path never sees these.
+"""
+
+LOCAL_EXTRAS_GPU_MESSAGES = {
+    'gpu_driver_issue_detected': [
+        "📋 DIAGNOSIS: NVIDIA Driver Issue Detected",
+        "",
+        "   CUDA toolkit appears to be installed on your system, but your NVIDIA GPU",
+        "   hardware is not accessible. This typically indicates a driver problem.",
+        "",
+        "   Possible causes:",
+        "      • NVIDIA drivers are missing, outdated, or corrupted",
+        "      • Driver installation was incomplete",
+        "      • Windows Update overwrote NVIDIA drivers with generic ones",
+        "      • GPU hardware is present but drivers can't communicate with it",
+        "",
+        "💡 This is usually fixable with a proper driver reinstall.",
+    ],
+    'gpu_hardware_missing': [
+        "📋 DIAGNOSIS: NVIDIA GPU Not Found",
+        "",
+        "   No NVIDIA GPU detected and no CUDA components found on this system.",
+        "",
+        "   Possible causes:",
+        "      • This computer does not have an NVIDIA GPU",
+        "      • NVIDIA GPU is disabled in BIOS/UEFI settings",
+        "      • GPU hardware is not properly connected",
+        "      • No NVIDIA software has ever been installed",
+        "",
+        "💡 Verify your hardware meets the requirements before proceeding.",
+    ],
+    'gpu_requirement_explanation': [
+        "📋 NVIDIA GPU REQUIREMENT:",
+        "",
+        "   This Monster Hunter game requires an NVIDIA GPU for optimal performance.",
+        "   The AI models use GPU acceleration for real-time inference during gameplay.",
+        "",
+        "      ⚡ With NVIDIA GPU: 1-2 second response times",
+        "      🐌 Without GPU (CPU-only): 30-60 second response times (35x slower)",
+        "",
+        "   Minimum requirements:",
+        "      • NVIDIA GPU with compute capability 6.0+ (GTX 1060 or newer)",
+        "      • 6GB+ VRAM recommended (4GB minimum)",
+        "      • Latest NVIDIA drivers",
+        "      • CUDA 11.8+ or CUDA 12.x toolkit",
+        "",
+        "   The game is designed for GPU acceleration and becomes practically",
+        "   unplayable without it due to the real-time AI interaction requirements.",
+    ],
+    'gpu_driver_troubleshooting': [
+        "📋 NVIDIA DRIVER TROUBLESHOOTING:",
+        "",
+        "      Step 1 - Download Latest Drivers:",
+        "      1. Go to https://www.nvidia.com/drivers/",
+        "      2. Enter your GPU model or let it auto-detect",
+        "      3. Download the latest Game Ready Driver (not Studio)",
+        "      4. Choose 'Custom Installation' when running installer",
+        "",
+        "   Step 2 - Clean Driver Installation:",
+        "      1. Download DDU (Display Driver Uninstaller) from guru3d.com",
+        "      2. Boot into Safe Mode (Advanced Startup → Troubleshoot → Advanced → Startup Settings)",
+        "      3. Run DDU and select 'Clean and Restart' for NVIDIA drivers",
+        "      4. After restart, install the new NVIDIA driver you downloaded",
+        "      5. Restart again after installation completes",
+        "",
+        "   Step 3 - Verification:",
+        "      1. Open Command Prompt",
+        "      2. Run: nvidia-smi",
+        "      3. You should see your GPU listed with temperature, memory usage",
+        "",
+        "   If issues persist:",
+        "      • Check Windows Update didn't reinstall generic drivers",
+        "      • Disable Windows automatic driver updates",
+        "      • Check BIOS settings (disable Secure Boot if enabled)",
+        "      • Verify GPU power connections and seating",
+        "",
+        "💡 Driver issues are the #1 cause of 'GPU not detected' problems",
+    ],
+    'gpu_hardware_troubleshooting': [
+        "📋 GPU HARDWARE TROUBLESHOOTING:",
+        "",
+        "   Step 1 - Verify Hardware Presence:",
+        "      1. Press Win+X, select 'Device Manager'",
+        "      2. Expand 'Display adapters'",
+        "      3. Look for your NVIDIA GPU (GeForce, RTX, GTX, etc.)",
+        "      4. If you see 'Microsoft Basic Display Adapter' - drivers missing",
+        "      5. If no NVIDIA GPU listed - hardware issue",
+        "",
+        "   Step 2 - Physical Hardware Check:",
+        "      1. Power off computer completely",
+        "      2. Check GPU is fully seated in PCIe slot",
+        "      3. Check all GPU power connectors (6-pin, 8-pin) are firmly connected",
+        "      4. Check PSU has enough wattage for your GPU",
+        "      5. Try reseating the GPU in the PCIe slot",
+        "",
+        "   Step 3 - BIOS/UEFI Settings:",
+        "      1. Boot into BIOS/UEFI (usually F2, F12, or Del during startup)",
+        "      2. Look for 'Primary Display' or 'Graphics' settings",
+        "      3. Set primary graphics to 'PCIe' or 'Discrete' (not 'Integrated')",
+        "      4. Disable 'Secure Boot' if enabled (can interfere with drivers)",
+        "      5. Enable 'Above 4G Decoding' if available",
+        "      6. Save settings and exit",
+        "",
+        "   Step 4 - Install NVIDIA Drivers:",
+        "      1. Boot into Windows",
+        "      2. Go to https://www.nvidia.com/drivers/",
+        "      3. Download and install latest drivers",
+        "      4. Restart computer",
+        "      5. Test with 'nvidia-smi' command",
+        "",
+        "   If still no GPU detected:",
+        "      • GPU may be faulty (try in another computer if possible)",
+        "      • Motherboard PCIe slot may be damaged",
+        "      • Power supply may be insufficient",
+        "      • GPU may not be compatible with this motherboard",
+        "",
+        "⚠️  If you don't have an NVIDIA GPU, this game requires one for good performance",
+    ],
+    'gpu_hardware_capable': [
+        "",
+        "✅ GPU HARDWARE CONFIRMED",
+        "",
+        "Your NVIDIA GPU meets the requirements for this AI game.",
+        "Expected minimum performance with your hardware:",
+        "",
+        "⚡  LLM inference: 20+ tokens per second",
+        "🖼️  Image generation: runs in the cloud (Gemini) - the GPU is not involved",
+        "🤖  The recommended AI models should work without any manual configuration.",
+        "",
+        "Any remaining setup issues are likely with software configuration rather than hardware limitations.",
+        "This means that they should be solvable since your hardware is confirmed working and capable.",
+    ],
+    'gpu_hardware_not_capable': [
+        "",
+        "⚠️  GPU CAPABILITY CONCERN",
+        "",
+        "Your NVIDIA GPU was detected but may struggle with this AI game's requirements.",
+        "",
+        "   What this means:",
+        "      • LLM responses may take longer than expected",
+        "      • Image generation may be very slow or fail",
+        "      • Gameplay experience may feel sluggish",
+        "      • Some AI features may not work reliably",
+        "",
+        "   Options to consider:",
+        "      1. Continue setup anyway - the game might still be playable",
+        "      2. Try with smaller/faster AI models (requires manual configuration)",
+        "      3. Consider upgrading to a more recent NVIDIA GPU",
+        "",
+        "   Minimum recommended: GTX 1060 6GB or RTX 2060",
+        "   Ideal for smooth gameplay: RTX 3070 or better",
+        "",
+        "The setup will continue, but performance expectations should be adjusted.",
+    ],
+    'gpu_driver_outdated': [
+        "",
+        "",
+        "📋 OUTDATED NVIDIA DRIVERS DETECTED",
+        "",
+        "   Your NVIDIA drivers are too old for modern CUDA development.",
+        "   This will prevent proper GPU acceleration and may cause compilation failures.",
+        "",
+        "   Minimum driver requirements:",
+        "      • CUDA 12.x support: Driver 530.30 or newer",
+        "      • CUDA 11.x support: Driver 470.57 or newer",
+        "      • For best compatibility: Latest Game Ready drivers",
+        "",
+        "💡 Driver updates are usually quick (5-10 minutes) and solve most issues",
+    ],
+    'gpu_driver_general_issues': [
+        "",
+        "",
+        "📋 NVIDIA DRIVER PROBLEMS DETECTED",
+        "",
+        "   There are issues with your NVIDIA driver installation.",
+        "   This will prevent proper GPU acceleration for AI workloads.",
+        "",
+        "   Common driver problems:",
+        "      • Drivers were corrupted during installation",
+        "      • Windows Update installed generic drivers over NVIDIA drivers",
+        "      • Multiple driver versions are conflicting",
+        "      • Driver installation was incomplete",
+        "      • System needs restart after driver changes",
+        "",
+        "   Impact on the game:",
+        "      • GPU acceleration may not work",
+        "      • CUDA compilation may fail",
+        "      • AI model inference will be very slow",
+        "",
+        "💡 Most driver issues can be fixed with a clean driver reinstall",
+        "",
+        "",
+    ],
+    'gpu_driver_fix_instructions': [
+        "",
+        "",
+        "📋 UPDATING NVIDIA DRIVERS:",
+        "",
+        "   Instructions:",
+        "     1. Go to https://www.nvidia.com/drivers/",
+        "     2. Select your GPU or use auto-detection",
+        "     3. Download latest Game Ready Driver (not Studio)",
+        "     4. Run installer and choose 'Express Installation'",
+        "     5. Restart computer when prompted",
+        "     6. Test with: nvidia-smi",
+        "",
+        "💡 After npdating Nvidia drivers, please restart this setup.",
+        "",
+        "",
+    ],
+    'cuda_toolkit_missing': [
+        "",
+        "",
+        "📋 CUDA TOOLKIT REQUIRED:",
+        "",
+        "   The CUDA Toolkit provides the development environment needed for GPU acceleration.",
+        "   Without it, AI models cannot use your NVIDIA GPU and will run extremely slowly on CPU.",
+        "",
+        "   What CUDA Toolkit provides:",
+        "      • CUDA compiler (nvcc) for building GPU-accelerated code",
+        "      • CUDA runtime libraries for executing GPU code",
+        "      • cuDNN libraries for deep learning acceleration",
+        "      • Development headers and documentation",
+        "",
+        "   Performance impact:",
+        "   ⚡ With CUDA: 1-2 second AI responses",
+        "   🐌 Without CUDA: 30-60 second AI responses (30x slower)",
+        "",
+        "💡 CUDA installation is required for this game to be playable",
+        "",
+    ],
+    'cuda_toolkit_installation': [
+        "",
+        "",
+        "📋 CUDA TOOLKIT INSTALLATION:",
+        "",
+        "   Download and install CUDA Toolkit (usually 2-3 GB): "
+        "   CUDA Toolkit: https://developer.nvidia.com/cuda-downloads",
+        "",
+        "   Verify Installation:",
+        "      1. Close all Command Prompt windows",
+        "      2. Open Command Prompt",
+        "      3. Run: nvcc --version",
+        "         - Should show CUDA compiler version (e.g., 'release 12.2')",
+        "      4. Run: nvidia-smi",
+        "         - Should show GPU info with CUDA version",
+        "",
+        "💡 After installing CUDA, please restart this setup.",
+        "",
+        "",
+    ],
+    'cuda_environment_issues': [
+        "",
+        "",
+        "📋 CUDA ENVIRONMENT CONFIGURATION:",
+        "",
+        "   CUDA toolkit appears to be installed, but environment configuration has issues.",
+        "   This prevents applications from finding and using CUDA properly.",
+        "",
+        "   Common environment problems:",
+        "      • CUDA compiler (nvcc) not in system PATH",
+        "      • CUDA_PATH environment variable not set",
+        "      • Multiple CUDA versions causing conflicts",
+        "      • Windows environment variables not updated after installation",
+        "      • Command prompt not restarted after CUDA installation",
+        "",
+        "   Why this matters:",
+        "      • Python packages like llama-cpp-python need to find CUDA during compilation",
+        "      • Without proper environment, GPU will be ignored even if CUDA is installed",
+        "",
+        "",
+    ],
+    'cuda_environment_fix': [
+        "",
+        "",
+        "📋 FIXING CUDA ENVIRONMENT CONFIGURATION:",
+        "",
+        "   Method 1 - Automatic Fix (Try First):",
+        "      1. Close all Command Prompt windows",
+        "      2. Restart your computer",
+        "      3. Open new Command Prompt",
+        "      4. Test commands:",
+        "            - nvcc --version",
+        "            - echo %CUDA_PATH%",
+        "      5. If this works, environment was just cached incorrectly",
+        "",
+        "   Method 2 - Manual Environment Variables:",
+        "      1. Find your CUDA installation directory:",
+        "         - Usually: C:\\Program Files\\NVIDIA GPU Computing Toolkit\\CUDA\\v12.x",
+        "         - Look for the newest version if multiple exist",
+        "",
+        "      2. Set environment variables:",
+        "         - Press Win+R, type 'sysdm.cpl', press Enter",
+        "         - Click 'Environment Variables'",
+        "         - In 'System Variables' section:",
+        "",
+        "      3. Add/Update CUDA_PATH:",
+        "         - Click 'New' (or select existing CUDA_PATH and click 'Edit')",
+        "         - Variable name: CUDA_PATH",
+        "         - Variable value: C:\\Program Files\\NVIDIA GPU Computing Toolkit\\CUDA\\v12.x",
+        "         - (Replace v12.x with your actual version)",
+        "",
+        "      4. Update PATH variable:",
+        "         - Find and select 'Path' in System Variables",
+        "         - Click 'Edit'",
+        "         - Add these paths (if not already present):",
+        "           * %CUDA_PATH%\\bin",
+        "           * %CUDA_PATH%\\libnvvp",
+        "         - Click 'OK' on all dialogs",
+        "",
+        "      5. Test the fix:",
+        "         - Close all Command Prompt windows",
+        "         - Open new Command Prompt",
+        "         - Run: nvcc --version",
+        "         - Run: echo %CUDA_PATH%",
+        "",
+        "   Method 3 - Reinstall CUDA (If Methods 1-2 Fail):",
+        "      1. Clean install:",
+        "         - Uninstall existing CUDA",
+        "         - Download latest CUDA toolkit",
+        "         - Restart computer when done",
+        "",
+    ],
+}
