+++ setup/messages/local_extras_llm.py
@@ -1,0 +1,426 @@
+"""
+Local-extras messages: Visual Studio Build Tools, llama-cpp-python, and
+GGUF model configuration - the toolchain that exists solely to compile
+and run the local LLM engine.
+Only shown by the --local-extras walkthrough (the unsupported local-LLM
+escape hatch) - a new player on the API-first path never sees these.
+"""
+
+LOCAL_EXTRAS_LLM_MESSAGES = {
+    'visual_studio_requirement_explanation': [
+        "",
+        "",
+        "📋 VISUAL STUDIO BUILD TOOLS REQUIREMENT:",
+        "",
+        "   This Monster Hunter game requires Visual Studio Build Tools to compile",
+        "   the llama-cpp-python package with CUDA support.",
+        "",
+        "   What needs to be installed:",
+        "      • Microsoft C++ compiler (MSVC)",
+        "      • Windows SDK for system headers",
+        "      • CMake and MSBuild tools",
+        "      • Total size: ~1-3 GB depending on options",
+        "",
+        "   SETUP:",
+        "      • The above checks may be false negative, ",
+        "      • If you believe you meet these requirements, choose [S] to skip this component",
+        "      • Otherwise, choose [G] to get further guidance",
+        "",
+        "💡 You only need 'Build Tools' - not the full Visual Studio IDE",
+        "",
+    ],
+    'visual_studio_installation_instructions': [
+        "",
+        "",
+        "📋 VISUAL STUDIO BUILD TOOLS INSTALLATION:",
+        "",
+        "   1. Go to https://visualstudio.microsoft.com/downloads/",
+        "",
+        "   2. Scroll down to 'All Downloads' section",
+        "",
+        "   3. Find and download 'Build Tools for Visual Studio 2022'",
+        "      (Look for the standalone 'Build Tools' - NOT the full IDE)",
+        "",
+        "   4. Run the downloaded installer (vs_buildtools.exe)",
+        "",
+        "   5. In the Visual Studio Installer, select workloads:",
+        "      •  Check 'C++ build tools' (this is the main workload you need)",
+        "",
+        "   6. In the 'Individual components' tab, ensure these are selected:",
+        "      •  MSVC v143 - VS 2022 C++ x64/x86 build tools (latest)",
+        "      •  Windows 11 SDK (10.0.22621.0) or Windows 10 SDK (latest)",
+        "      •  CMake tools for Visual Studio",
+        "",
+        "   7. Click 'Install' and wait for completion (10-30 minutes)",
+        "",
+        "   8. Restart your computer when installation finishes",
+        "",
+        "",
+    ],
+    'visual_studio_modify_for_cpp_tools': [
+        "",
+        "",
+        "📋 ADDING C++ TOOLS TO EXISTING VISUAL STUDIO:",
+        "",
+        "   You have Visual Studio installed but it's missing C++ compilation capabilities.",
+        "   Here's how to add them:",
+        "",
+        "   Method 1 - Visual Studio Installer (Recommended):",
+        "      1. Press Windows key and search for 'Visual Studio Installer'",
+        "      2. Launch the Visual Studio Installer",
+        "      3. Find your Visual Studio installation and click 'Modify'",
+        "      4. Go to the 'Workloads' tab",
+        "      5. Check the box for:",
+        "         • 'C++ build tools' (for Build Tools edition)",
+        "         OR",
+        "         • 'Desktop development with C++' (for Community/Professional edition)",
+        "",
+        "      6. Click 'Individual components' tab and verify these are selected:",
+        "         • MSVC v143 - VS 2022 C++ x64/x86 build tools",
+        "         • Windows 11 SDK or Windows 10 SDK (latest version)",
+        "         • CMake tools for Visual Studio",
+        "",
+        "      7. Click 'Modify' and wait for installation (5-20 minutes)",
+        "",
+        "      8. Restart your computer after installation completes",
+        "",
+        "   Method 2 - Fresh Install (If Modify Doesn't Work):",
+        "      1. Uninstall current Visual Studio from Windows Settings",
+        "      2. Download Build Tools for Visual Studio 2022",
+        "      3. Follow the fresh installation instructions",
+        "",
+        "💡 If you can't find Visual Studio Installer:",
+        "   - Download it again from https://visualstudio.microsoft.com/downloads/",
+        "   - Look for 'Visual Studio Installer' in the Tools section",
+        "",
+    ],
+    'visual_studio_missing_cpp_tools_detected': [
+        "",
+        "📋 VISUAL STUDIO MISSING C++ BUILD TOOLS:",
+        "",
+        "   Visual Studio is installed on your system, but it doesn't include C++ compilation capabilities.",
+        "   This is a common configuration when Visual Studio was installed for other development",
+        "   (like web development, .NET, or Python) without including C++ tools.",
+        "",
+        "   What you have:",
+        "      ✅ Visual Studio installation",
+        "      ✅ Windows SDK",
+        "      ❌ C++ build tools (MSVC compiler)",
+        "",
+        "   What you need:",
+        "      • Add the C++ build tools workload to your existing Visual Studio installation",
+        "",
+        "💡 This can be fixed by modifying your current Visual Studio installation",
+        "",
+    ],
+    'visual_studio_sdk_detection_issue': [
+        "",
+        "📋 WINDOWS SDK DETECTION ISSUE:",
+        "",
+        "   Visual Studio and C++ build tools are detected, but Windows SDK was not found",
+        "   in the expected locations. This is likely a false negative.",
+        "",
+        "   What you have:",
+        "      ✅ Visual Studio installation",
+        "      ✅ C++ build tools (MSVC compiler)",
+        "      ❌ Windows SDK (may be bundled with Visual Studio)",
+        "",
+        "   Most likely cause:",
+        "      • Windows SDK is included with your Visual Studio installation",
+        "      • SDK components are in VS directory instead of standalone location",
+        "",
+        "💡 Your build environment is probably functional despite this detection issue",
+        "",
+    ],
+    'visual_studio_partial_installation_detected': [
+        "",
+        "📋 PARTIAL DEVELOPMENT ENVIRONMENT DETECTED:",
+        "",
+        "   Some development components were found but Visual Studio is not installed",
+        "   in the expected locations. This suggests an incomplete or non-standard setup.",
+        "",
+        "   What was detected:",
+        "      ❌ Visual Studio installation (not found in standard directories)",
+        "      ✅ Some development components (Windows SDK or build tools)",
+        "",
+        "   Possible causes:",
+        "      • Visual Studio installed in non-standard location",
+        "      • Standalone Windows SDK installation without Visual Studio",
+        "      • Incomplete or corrupted Visual Studio installation",
+        "",
+        "💡 A complete Visual Studio Build Tools installation is recommended",
+        "",
+    ],
+    'llama_cpp_requirement_explanation': [
+        "",
+        "📋 LLM INTEGRATION REQUIREMENT:",
+        "",
+        "   This Monster Hunter game requires llama-cpp-python for AI language model inference.",
+        "   The game's monsters, NPCs, and story generation depend on fast AI responses.",
+        "",
+        "   Performance requirements:",
+        "      ⚡ Game designed for: ~35+ tokens/second (smooth gameplay)",
+        "      🐌 CPU-only typically: <3 tokens/second (game becomes unplayable)",
+        "",
+        "   What you're choosing:",
+        "      • CUDA version: Compiles with your NVIDIA GPU for fast inference",
+        "      • CPU version: Uses only your processor (extremely slow for gaming)",
+        "",
+        "   Installation complexity:",
+        "      • CUDA installation tries 3 different methods automatically",
+        "      • May take 5-15 minutes depending on your system",
+        "      • Requires Visual Studio Build Tools and CUDA toolkit",
+        "",
+        "💡 CUDA acceleration is essential for playable game performance",
+        "",
+    ],
+    'llama_cpp_cuda_install_failed': [
+        "",
+        "📋 CUDA INSTALLATION FAILED:",
+        "",
+        "   The automatic CUDA installation encountered problems.",
+        "   This usually indicates missing prerequisites or system configuration issues.",
+        "",
+        "   Common causes:",
+        "      • Visual Studio Build Tools not properly installed",
+        "      • CUDA toolkit missing or not in PATH",
+        "      • NVIDIA drivers too old or missing",
+        "      • Network issues downloading packages",
+        "      • Insufficient disk space for compilation",
+        "",
+        "   What you can do:",
+        "      1. Retry installation (sometimes temporary issues resolve)",
+        "      2. Install CPU-only version (playable but very slow)",
+        "      3. Skip for now and troubleshoot prerequisites",
+        "",
+        "💡 CPU-only version lets you test the game while fixing CUDA issues",
+        "",
+    ],
+    'llama_cpp_cpu_warning': [
+        "",
+        "⚠️  CPU-ONLY INSTALLATION WARNING:",
+        "",
+        "   You're about to install the CPU-only version of llama-cpp-python.",
+        "   This will make the game extremely slow and frustrating to play.",
+        "",
+        "   Expected performance impact:",
+        "      • Token generation: Typically <3 tokens/second (vs 35+ needed)",
+        "      • Game timeouts: Prompts will frequently timeout and retry",
+        "      • Gameplay flow: Constant interruptions from slow generation",
+        "      • Overall experience: Nearly unplayable due to performance issues",
+        "",
+        "   Technical issues with slow generation:",
+        "      • Game assumes broken AI when prompts take too long",
+        "      • Automatic retries can create endless loops",
+        "      • Even with adjusted timeout settings, gameplay remains unbearably slow",
+        "",
+        "   Why continue with CPU-only:",
+        "      ✓ Test that the game installation works",
+        "      ✓ Verify other components before fixing GPU acceleration",
+        "      ✓ Temporary solution while troubleshooting CUDA",
+        "",
+        "   You can upgrade to CUDA later by running setup again.",
+        "",
+        "💡 This is only recommended for testing, not actual gameplay",
+        "",
+    ],
+    'llama_cpp_broken_installation': [
+        "",
+        "📋 BROKEN INSTALLATION DETECTED:",
+        "",
+        "   llama-cpp-python is installed but cannot be imported properly.",
+        "   This usually indicates a corrupted or incomplete installation.",
+        "",
+        "   Common causes:",
+        "      • Installation was interrupted",
+        "      • Mixed CUDA and CPU versions installed",
+        "      • Missing dependencies or DLLs",
+        "      • Virtual environment corruption",
+        "      • Python version compatibility issues",
+        "",
+        "   Solution:",
+        "      • Clean reinstallation is the most reliable fix",
+        "      • This will remove the broken installation and start fresh",
+        "",
+        "💡 Reinstallation usually resolves import and compatibility issues",
+        "",
+    ],
+    'llama_cpp_cpu_detected': [
+        "",
+        "📋 CPU-ONLY PERFORMANCE DETECTED:",
+        "",
+        "   Performance test indicates llama-cpp-python is running on CPU only.",
+        "   This will make the game extremely slow and potentially broken.",
+        "",
+        "   Current performance:",
+        "      • Very slow token generation (under 3 tokens/second)",
+        "      • Game designed for 35+ tokens/second performance",
+        "      • Prompts will frequently timeout and retry",
+        "      • Game may become stuck in retry loops",
+        "",
+        "   Technical issues:",
+        "      • Game assumes AI is broken when generation is too slow",
+        "      • Timeout settings can be adjusted but don't solve core speed issue",
+        "      • Even with maximum timeouts, gameplay becomes unbearably slow",
+        "",
+        "   Why this happened:",
+        "      • CUDA installation failed silently",
+        "      • GPU drivers not working properly",
+        "      • CUDA libraries not found at runtime",
+        "      • CPU-only version was installed intentionally",
+        "",
+        "   Recommendations:",
+        "      • Upgrade to CUDA version for playable performance",
+        "      • Keep CPU version only for testing purposes",
+        "",
+        "💡 The game is designed for GPU acceleration and needs it to be fun",
+        "",
+    ],
+    'llama_cpp_weak_gpu_detected': [
+        "",
+        "📋 WEAK GPU PERFORMANCE DETECTED:",
+        "",
+        "   Performance test shows slower than expected GPU acceleration.",
+        "   The game should still be playable but may feel sluggish at times.",
+        "",
+        "   Current performance:",
+        "      • Moderate token generation (3-10 tokens/second)",
+        "      • Below optimal 35+ tokens/second target",
+        "      • Prompts will frequently timeout and retry",
+        "      • Game may become stuck in retry loops",
+        "",
+        "   Possible causes:",
+        "      • Older or entry-level GPU",
+        "      • Not all model layers loaded on GPU",
+        "      • GPU memory limitations",
+        "      • Suboptimal CUDA installation",
+        "",
+        "   Options:",
+        "      • Try CUDA reinstall (may improve GPU utilization)",
+        "      • Install a smaller faster model (quality will drop)",
+        "      • Adjust model settings later for better performance",
+        "",
+        "💡 You may need to adjust timeout limits in lmm config",
+        "",
+    ],
+    'llm_model_requirement_explanation': [
+        "",
+        "📋 LLM MODEL CONFIGURATION REQUIRED:",
+        "",
+        "   This Monster Hunter game requires a language model file for AI-powered",
+        "   monster interactions, NPC conversations, and dynamic story generation.",
+        "",
+        "   What you need:",
+        "      • A GGUF format language model (recommended: 7B or 13B parameters)",
+        "      • Model file stored locally on your computer",
+        "      • At least 8GB free disk space for the model",
+        "      • Model configured in the .env file for the game to find it",
+        "",
+        "   Recommended Model: Kunoichi 7B",
+        "",
+        "   Other popular model options:",
+        "      • Llama 2 7B Chat (good balance of speed and quality)",
+        "      • Mistral 7B Instruct (fast and capable)",
+        "      • CodeLlama 7B Instruct (good for technical content)",
+        "",
+        "   Where to get models:",
+        "      • Hugging Face (search for GGUF format)",
+        "      • TheBloke's model collection (pre-quantized GGUF)",
+        "      • Local model conversion tools",
+        "",
+        "💡 The game is designed to work with most 7B+ parameter models in GGUF format",
+        "",
+    ],
+    'llm_model_path_placeholder': [
+        "",
+        "📋 SETTING LLM MODEL PATH:",
+        "",
+        "   Please enter the full path to your model file",
+        "",
+        "   Example valid paths:",
+        "      C:/Users/YourName/.cache/lm-studio/models/TheBloke/Kunoichi-7B-GGUF/kunoichi-7b.Q6_K.gguf",
+        "      D:/Downloads/mistral-7b-instruct-v0.1.Q4_K_M.gguf",
+        "      C:/AI/models/llama-2-7b-chat.q4_K_M.gguf",
+        "",
+        "💡 Use forward slashes (/) in the path, even on Windows",
+        "",
+    ],
+    'llm_model_file_missing': [
+        "",
+        "📋 MODEL FILE NOT FOUND:",
+        "",
+        "   Your .env file has a model path configured, but the file doesn't exist",
+        "   at that location. The file may have been moved, deleted, or the path",
+        "   may contain a typo.",
+        "",
+        "   Common solutions:",
+        "      • Check if the file was moved to a different folder",
+        "      • Verify the filename is spelled correctly (including extension)",
+        "      • Make sure the drive letter is correct (C:/, D:/, etc.)",
+        "      • Check if the file was accidentally deleted",
+        "",
+        "   Model file requirements:",
+        "      • Must be in GGUF format (.gguf extension)",
+        "      • Must be at least 100MB in size",
+        "      • Must be accessible/readable by the game",
+        "",
+        "💡 You can download models from Hugging Face or TheBloke's collection",
+        "",
+    ],
+    'llm_model_path_invalid': [
+        "",
+        "📋 INVALID MODEL PATH CONFIGURATION:",
+        "",
+        "   There's an issue with your current model path configuration.",
+        "   This could be due to various reasons like incorrect file type,",
+        "   permission issues, or file corruption.",
+        "",
+        "   Common issues:",
+        "      • File is not a valid model format (.gguf, .ggml, .bin, .safetensors)",
+        "      • File is too small to be a language model (< 100MB)",
+        "      • Permission denied when trying to access the file",
+        "      • File appears to be corrupted or incomplete",
+        "",
+        "   What to check:",
+        "      • File extension should be .gguf for best compatibility",
+        "      • File size should be several GB for good models",
+        "      • File should be in a location you have read access to",
+        "",
+        "💡 GGUF format models from Hugging Face are most reliable",
+        "",
+    ],
+    'llm_model_download_guidance': [
+        "",
+        "📋 HOW TO FIND AND DOWNLOAD A LANGUAGE MODEL:",
+        "",
+        "   Method 1 - Hugging Face (Direct Download):",
+        "      • Go to https://huggingface.co/",
+        "      • Search for models by 'TheBloke' (pre-quantized GGUF models)",
+        "      • Good starting models: 'Llama-2-7B-Chat-GGUF', 'Mistral-7B-Instruct-v0.1-GGUF'",
+        "      • Click model → 'Files and versions' tab → download the .gguf file",
+        "      • Choose Q4_K_M version for best balance of quality and speed",
+        "",
+        "   Method 2 - LM Studio (Easiest for Beginners):",
+        "      • Download LM Studio from https://lmstudio.ai/",
+        "      • Install and open LM Studio",
+        "      • Go to 'Discover' tab and search for models",
+        "      • Download any model you like (LM Studio handles GGUF automatically)",
+        "      • Models are saved to: C:/Users/YourName/.cache/lm-studio/models/",
+        "",
+        "   Understanding Quantization (Q4_K_M, Q5_K_M, etc.):",
+        "      • Quantization = compression technique to reduce model size",
+        "      • Q4_K_M: ~4GB, good quality, fast (recommended for gaming)",
+        "      • Q5_K_M: ~5GB, better quality, slower",
+        "      • Q8_0: ~7GB, highest quality, slowest",
+        "      • Start with Q4_K_M unless you have plenty of VRAM",
+        "",
+        "   File Management:",
+        "      • Save models to a dedicated folder like C:/AI/models/ or D:/models/",
+        "      • Models are typically 3-8GB each, so ensure you have disk space",
+        "      • Use descriptive folder names if you plan to try multiple models",
+        "      • The exact folder location doesn't matter as long as you remember it",
+        "",
+        "💡 LM Studio is the easiest way to get started - it handles everything automatically",
+        "",
+    ],
+}
