--- setup/user_messages.py
@@ -1,1020 +1,0 @@
-# Installation Instructions
-# Centralized repository of all verbose installation instructions
-# Used by setup modules to keep code clean and instructions consistent
-
-MESSAGES = {
-    'nodejs_installation': [
-        "📋 INSTALLATION REQUIRED:",
-        "   1. Go to https://nodejs.org/",
-        "   2. Download the LTS version (version 18 or later)",
-        "   3. Run the downloaded installer",
-        "      - Accept all default settings unless you know what you're doing",
-        "",
-        "   4. After installation, open a terminal and run the following commands to verify:",
-        "      - `node -v` (should return a version like v18.x.x or higher)",
-        "      - `npm -v` (should return a version number)",
-        "",
-        "   5. Once installed, please re-run start_game.bat",
-        "",
-        "💡 This usually takes 2–3 minutes to download and install.",
-        "💡 You'll need to restart your command prompt after installation.",
-        "",
-        "You have the following options to continue:",
-        "   1. Skip and move on to next component (recommended)",
-        "   2. Try to install frontend dependencies anyways (if nodejs is installed, but not detected)",
-        "   3. Exit the setup and restart when nodejs is installed",
-        "",
-    ],
-    'mysql_installation': [
-        "📋 INSTALLING MYSQL - WHAT TO DO:",
-        "",
-        "The MySQL download page just opened in your browser.",
-        "(If it didn't, go to: https://dev.mysql.com/downloads/mysql/)",
-        "",
-        "1. Download the INSTALLER for Windows (the MSI download, not the ZIP)",
-        "   and run the file it gives you.",
-        "",
-        "2. The wizard asks a lot of questions. Only these answers matter:",
-        "",
-        "     Setup type ...... 'Server only' is plenty",
-        "     Port ............ keep 3306",
-        "     Root password ... MAKE ONE UP and keep it handy -",
-        "                       this setup asks for it in a minute",
-        "     Windows Service . yes, and start at system startup",
-        "",
-        "   For everything else, accept what the wizard suggests.",
-        "",
-        "3. Let it finish (5-10 minutes). If a separate 'Configurator'",
-        "   window appears near the end, finish that too - it's the part",
-        "   that actually starts the server.",
-        "",
-        "This setup will wait here and check as soon as you're done.",
-        "",
-    ],
-    'mysql_service_start': [
-        "📋 STARTING MYSQL BY HAND:",
-        "",
-        "1. Press the Windows key + R together, type  services.msc",
-        "   and press Enter - a window called 'Services' opens.",
-        "2. Scroll to the entries starting with 'MySQL' (like MySQL84).",
-        "3. Right-click it and choose 'Start'.",
-        "4. So it starts by itself from now on: right-click it again,",
-        "   choose 'Properties', set 'Startup type' to 'Automatic', OK.",
-        "",
-        "If Windows says it can't start the service, a reinstall with the",
-        "wizard's suggested answers is usually faster than debugging it.",
-    ],
-    'mysql_troubleshooting': [
-        "📋 MYSQL TROUBLESHOOTING:",
-        "",
-        "Common issues and solutions:",
-        "",
-        "Issue 1 - Port 3306 in use:",
-        "1. Open Command Prompt as Administrator",
-        "2. Run: netstat -ano | findstr :3306",
-        "3. If another process is using port 3306:",
-        "   - Stop that process, or",
-        "   - Reconfigure MySQL to use a different port",
-        "",
-        "Issue 2 - MySQL won't start:",
-        "1. Check Windows Event Viewer:",
-        "   - Press Win+R → eventvwr.msc",
-        "   - Go to Windows Logs → Application",
-        "   - Look for MySQL errors",
-        "2. Common fixes:",
-        "   - Delete ib_logfile0 and ib_logfile1 from MySQL data directory",
-        "   - Run MySQL as Administrator",
-        "   - Check disk space (MySQL needs space for logs)",
-        "",
-        "Issue 3 - Permission errors:",
-        "1. Run Command Prompt as Administrator",
-        "2. Navigate to MySQL bin directory",
-        "3. Try: mysqld --install",
-        "4. Then: net start mysql",
-        "",
-        "Issue 4 - Configuration problems:",
-        "1. Locate my.ini file (usually in MySQL installation folder)",
-        "2. Check for syntax errors",
-        "3. Reset to default configuration if needed",
-        "",
-        "Last resort:",
-        "- Completely uninstall MySQL",
-        "- Delete all MySQL folders",
-        "- Reinstall with default settings",
-    ],
-    'database_password_setup': [
-        "📋 YOUR MYSQL PASSWORD:",
-        "",
-        "The game needs the MySQL password to store your monsters and",
-        "progress. It's the one you made up during the MySQL install",
-        "(the 'root password').",
-        "",
-        "You'll only be asked once - it's saved so the game can use it.",
-        "",
-    ],
-    'database_connection_troubleshooting': [
-        "📋 START MYSQL SERVER:",
-        "",
-        "Your MySQL server needs to be running for the game to work.",
-        "Here's how to start it:",
-        "",
-        "Method 1 - Windows Services (Recommended):",
-        "1. Press Win+R, type 'services.msc', press Enter",
-        "2. Look for 'MySQL80' or 'MySQL84' or 'MySQL'",
-        "3. If status shows 'Stopped', right-click → 'Start'",
-        "4. Status should change to 'Running'",
-        "",
-        "Method 2 - Command Line:",
-        "1. Open Command Prompt as Administrator",
-        "2. Try: net start MySQL80",
-        "3. Or try: net start MySQL84",
-        "4. Or try: net start MySQL",
-        "5. You should see: 'The MySQL service was started successfully'",
-        "",
-        "Method 3 - MySQL Workbench:",
-        "1. Open MySQL Workbench (if installed)",
-        "2. Look for 'Server Status' or 'Server Administration'",
-        "3. Click 'Start Server'",
-        "",
-        "💡 MySQL server must be running every time you play the game",
-    ],
-    'database_troubleshooting': [
-        "📋 GENERAL DATABASE TROUBLESHOOTING:",
-        "",
-        "Connection issues can have several causes:",
-        "",
-        "Issue 1 - Firewall blocking connection:",
-        "1. Check Windows Firewall settings",
-        "2. Allow MySQL through firewall (port 3306)",
-        "3. Or temporarily disable firewall to test",
-        "",
-        "Issue 2 - MySQL configuration problems:",
-        "1. Find my.ini file (usually in MySQL installation folder)",
-        "2. Check bind-address setting (should be 0.0.0.0 or commented)",
-        "3. Restart MySQL service after changes",
-        "",
-        "Issue 3 - User permissions:",
-        "1. Connect to MySQL as root",
-        "2. Check user permissions: SHOW GRANTS FOR 'root'@'localhost';",
-        "3. Grant permissions if needed: GRANT ALL PRIVILEGES ON *.* TO 'root'@'localhost';",
-        "",
-        "Issue 4 - Wrong host/port in .env:",
-        "1. Check .env file has correct settings:",
-        "   DB_HOST=localhost",
-        "   DB_PORT=3306",
-        "2. Verify MySQL is actually running on these",
-        "",
-        "Issue 5 - MySQL version compatibility:",
-        "1. Check MySQL version: mysql --version",
-        "2. Older versions may have different authentication",
-        "3. Try: ALTER USER 'root'@'localhost' IDENTIFIED WITH mysql_native_password BY 'your_password';",
-        "",
-        "Last resort - Reinstall MySQL:",
-        "1. Uninstall MySQL completely",
-        "2. Delete all MySQL folders and data",
-        "3. Reinstall with default settings",
-    ],
-    'database_manual_creation': [
-        "📋 MANUAL DATABASE CREATION:",
-        "",
-        "   Create the game database manually using one of these methods:",
-        "",
-        "   Method 1 - MySQL Workbench (Easiest):",
-        "      1. Open MySQL Workbench",
-        "      2. Connect to your local MySQL server",
-        "      3. In the query tab, run:",
-        "         CREATE DATABASE IF NOT EXISTS monster_hunter_game;",
-        "      4. Click the lightning bolt to execute",
-        "      5. Verify: You should see 'monster_hunter_game' in the schemas panel",
-        "",
-        "   Method 2 - Command Line:",
-        "      . Open Command Prompt",
-        "      2. Connect to MySQL:",
-        "         mysql -u root -p",
-        "      3. Enter your MySQL root password",
-        "      4. Create database:",
-        "         CREATE DATABASE IF NOT EXISTS monster_hunter_game;",
-        "      5. Verify:",
-        "         SHOW DATABASES;",
-        "         (You should see 'monster_hunter_game' in the list)",
-        "      6. Exit: EXIT;",
-        "",
-        "   Method 3 - phpMyAdmin (if you have XAMPP/WAMP):",
-        "      1. Open phpMyAdmin in browser",
-        "      2. Click 'Databases' tab",
-        "      3. Enter 'monster_hunter_game' as database name",
-        "      4. Click 'Create'",
-        "",
-        "   To test the database was created:",
-        "      1. Run this setup again",
-        "      2. Or manually test: mysql -u root -p -e \"USE monster_hunter_game; SELECT 1;\"",
-        "",
-        "💡 Database name must be exactly: monster_hunter_game",
-    ],
-    'gpu_driver_issue_detected': [
-        "📋 DIAGNOSIS: NVIDIA Driver Issue Detected",
-        "",
-        "   CUDA toolkit appears to be installed on your system, but your NVIDIA GPU",
-        "   hardware is not accessible. This typically indicates a driver problem.",
-        "",
-        "   Possible causes:",
-        "      • NVIDIA drivers are missing, outdated, or corrupted",
-        "      • Driver installation was incomplete",
-        "      • Windows Update overwrote NVIDIA drivers with generic ones",
-        "      • GPU hardware is present but drivers can't communicate with it",
-        "",
-        "💡 This is usually fixable with a proper driver reinstall.",
-    ],
-    'gpu_hardware_missing': [
-        "📋 DIAGNOSIS: NVIDIA GPU Not Found",
-        "",
-        "   No NVIDIA GPU detected and no CUDA components found on this system.",
-        "",
-        "   Possible causes:",
-        "      • This computer does not have an NVIDIA GPU",
-        "      • NVIDIA GPU is disabled in BIOS/UEFI settings",
-        "      • GPU hardware is not properly connected",
-        "      • No NVIDIA software has ever been installed",
-        "",
-        "💡 Verify your hardware meets the requirements before proceeding.",
-    ],
-    'gpu_requirement_explanation': [
-        "📋 NVIDIA GPU REQUIREMENT:",
-        "",
-        "   This Monster Hunter game requires an NVIDIA GPU for optimal performance.",
-        "   The AI models use GPU acceleration for real-time inference during gameplay.",
-        "",
-        "      ⚡ With NVIDIA GPU: 1-2 second response times",
-        "      🐌 Without GPU (CPU-only): 30-60 second response times (35x slower)",
-        "",
-        "   Minimum requirements:",
-        "      • NVIDIA GPU with compute capability 6.0+ (GTX 1060 or newer)",
-        "      • 6GB+ VRAM recommended (4GB minimum)",
-        "      • Latest NVIDIA drivers",
-        "      • CUDA 11.8+ or CUDA 12.x toolkit",
-        "",
-        "   The game is designed for GPU acceleration and becomes practically",
-        "   unplayable without it due to the real-time AI interaction requirements.",
-    ],
-    'gpu_driver_troubleshooting': [
-        "📋 NVIDIA DRIVER TROUBLESHOOTING:",
-        "",
-        "      Step 1 - Download Latest Drivers:",
-        "      1. Go to https://www.nvidia.com/drivers/",
-        "      2. Enter your GPU model or let it auto-detect",
-        "      3. Download the latest Game Ready Driver (not Studio)",
-        "      4. Choose 'Custom Installation' when running installer",
-        "",
-        "   Step 2 - Clean Driver Installation:",
-        "      1. Download DDU (Display Driver Uninstaller) from guru3d.com",
-        "      2. Boot into Safe Mode (Advanced Startup → Troubleshoot → Advanced → Startup Settings)",
-        "      3. Run DDU and select 'Clean and Restart' for NVIDIA drivers",
-        "      4. After restart, install the new NVIDIA driver you downloaded",
-        "      5. Restart again after installation completes",
-        "",
-        "   Step 3 - Verification:",
-        "      1. Open Command Prompt",
-        "      2. Run: nvidia-smi",
-        "      3. You should see your GPU listed with temperature, memory usage",
-        "",
-        "   If issues persist:",
-        "      • Check Windows Update didn't reinstall generic drivers",
-        "      • Disable Windows automatic driver updates",
-        "      • Check BIOS settings (disable Secure Boot if enabled)",
-        "      • Verify GPU power connections and seating",
-        "",
-        "💡 Driver issues are the #1 cause of 'GPU not detected' problems",
-    ],
-    'gpu_hardware_troubleshooting': [
-        "📋 GPU HARDWARE TROUBLESHOOTING:",
-        "",
-        "   Step 1 - Verify Hardware Presence:",
-        "      1. Press Win+X, select 'Device Manager'",
-        "      2. Expand 'Display adapters'",
-        "      3. Look for your NVIDIA GPU (GeForce, RTX, GTX, etc.)",
-        "      4. If you see 'Microsoft Basic Display Adapter' - drivers missing",
-        "      5. If no NVIDIA GPU listed - hardware issue",
-        "",
-        "   Step 2 - Physical Hardware Check:",
-        "      1. Power off computer completely",
-        "      2. Check GPU is fully seated in PCIe slot",
-        "      3. Check all GPU power connectors (6-pin, 8-pin) are firmly connected",
-        "      4. Check PSU has enough wattage for your GPU",
-        "      5. Try reseating the GPU in the PCIe slot",
-        "",
-        "   Step 3 - BIOS/UEFI Settings:",
-        "      1. Boot into BIOS/UEFI (usually F2, F12, or Del during startup)",
-        "      2. Look for 'Primary Display' or 'Graphics' settings",
-        "      3. Set primary graphics to 'PCIe' or 'Discrete' (not 'Integrated')",
-        "      4. Disable 'Secure Boot' if enabled (can interfere with drivers)",
-        "      5. Enable 'Above 4G Decoding' if available",
-        "      6. Save settings and exit",
-        "",
-        "   Step 4 - Install NVIDIA Drivers:",
-        "      1. Boot into Windows",
-        "      2. Go to https://www.nvidia.com/drivers/",
-        "      3. Download and install latest drivers",
-        "      4. Restart computer",
-        "      5. Test with 'nvidia-smi' command",
-        "",
-        "   If still no GPU detected:",
-        "      • GPU may be faulty (try in another computer if possible)",
-        "      • Motherboard PCIe slot may be damaged",
-        "      • Power supply may be insufficient",
-        "      • GPU may not be compatible with this motherboard",
-        "",
-        "⚠️  If you don't have an NVIDIA GPU, this game requires one for good performance",
-    ],
-    'gpu_hardware_capable': [
-        "",
-        "✅ GPU HARDWARE CONFIRMED",
-        "",
-        "Your NVIDIA GPU meets the requirements for this AI game.",
-        "Expected minimum performance with your hardware:",
-        "",
-        "⚡  LLM inference: 20+ tokens per second",
-        "🖼️  Image generation: runs in the cloud (Gemini) - the GPU is not involved",
-        "🤖  The recommended AI models should work without any manual configuration.",
-        "",
-        "Any remaining setup issues are likely with software configuration rather than hardware limitations.",
-        "This means that they should be solvable since your hardware is confirmed working and capable.",
-    ],
-    'gpu_hardware_not_capable': [
-        "",
-        "⚠️  GPU CAPABILITY CONCERN",
-        "",
-        "Your NVIDIA GPU was detected but may struggle with this AI game's requirements.",
-        "",
-        "   What this means:",
-        "      • LLM responses may take longer than expected",
-        "      • Image generation may be very slow or fail",
-        "      • Gameplay experience may feel sluggish",
-        "      • Some AI features may not work reliably",
-        "",
-        "   Options to consider:",
-        "      1. Continue setup anyway - the game might still be playable",
-        "      2. Try with smaller/faster AI models (requires manual configuration)",
-        "      3. Consider upgrading to a more recent NVIDIA GPU",
-        "",
-        "   Minimum recommended: GTX 1060 6GB or RTX 2060",
-        "   Ideal for smooth gameplay: RTX 3070 or better",
-        "",
-        "The setup will continue, but performance expectations should be adjusted.",
-    ],
-    'gpu_driver_outdated': [
-        "",
-        "",
-        "📋 OUTDATED NVIDIA DRIVERS DETECTED",
-        "",
-        "   Your NVIDIA drivers are too old for modern CUDA development.",
-        "   This will prevent proper GPU acceleration and may cause compilation failures.",
-        "",
-        "   Minimum driver requirements:",
-        "      • CUDA 12.x support: Driver 530.30 or newer",
-        "      • CUDA 11.x support: Driver 470.57 or newer",
-        "      • For best compatibility: Latest Game Ready drivers",
-        "",
-        "💡 Driver updates are usually quick (5-10 minutes) and solve most issues",
-    ],
-    'gpu_driver_general_issues': [
-        "",
-        "",
-        "📋 NVIDIA DRIVER PROBLEMS DETECTED",
-        "",
-        "   There are issues with your NVIDIA driver installation.",
-        "   This will prevent proper GPU acceleration for AI workloads.",
-        "",
-        "   Common driver problems:",
-        "      • Drivers were corrupted during installation",
-        "      • Windows Update installed generic drivers over NVIDIA drivers",
-        "      • Multiple driver versions are conflicting",
-        "      • Driver installation was incomplete",
-        "      • System needs restart after driver changes",
-        "",
-        "   Impact on the game:",
-        "      • GPU acceleration may not work",
-        "      • CUDA compilation may fail",
-        "      • AI model inference will be very slow",
-        "",
-        "💡 Most driver issues can be fixed with a clean driver reinstall",
-        "",
-        "",
-    ],
-    'gpu_driver_fix_instructions': [
-        "",
-        "",
-        "📋 UPDATING NVIDIA DRIVERS:",
-        "",
-        "   Instructions:",
-        "     1. Go to https://www.nvidia.com/drivers/",
-        "     2. Select your GPU or use auto-detection",
-        "     3. Download latest Game Ready Driver (not Studio)",
-        "     4. Run installer and choose 'Express Installation'",
-        "     5. Restart computer when prompted",
-        "     6. Test with: nvidia-smi",
-        "",
-        "💡 After npdating Nvidia drivers, please restart this setup.",
-        "",
-        "",
-    ],
-    'cuda_toolkit_missing': [
-        "",
-        "",
-        "📋 CUDA TOOLKIT REQUIRED:",
-        "",
-        "   The CUDA Toolkit provides the development environment needed for GPU acceleration.",
-        "   Without it, AI models cannot use your NVIDIA GPU and will run extremely slowly on CPU.",
-        "",
-        "   What CUDA Toolkit provides:",
-        "      • CUDA compiler (nvcc) for building GPU-accelerated code",
-        "      • CUDA runtime libraries for executing GPU code",
-        "      • cuDNN libraries for deep learning acceleration",
-        "      • Development headers and documentation",
-        "",
-        "   Performance impact:",
-        "   ⚡ With CUDA: 1-2 second AI responses",
-        "   🐌 Without CUDA: 30-60 second AI responses (30x slower)",
-        "",
-        "💡 CUDA installation is required for this game to be playable",
-        "",
-    ],
-    'cuda_toolkit_installation': [
-        "",
-        "",
-        "📋 CUDA TOOLKIT INSTALLATION:",
-        "",
-        "   Download and install CUDA Toolkit (usually 2-3 GB): "
-        "   CUDA Toolkit: https://developer.nvidia.com/cuda-downloads",
-        "",
-        "   Verify Installation:",
-        "      1. Close all Command Prompt windows",
-        "      2. Open Command Prompt",
-        "      3. Run: nvcc --version",
-        "         - Should show CUDA compiler version (e.g., 'release 12.2')",
-        "      4. Run: nvidia-smi",
-        "         - Should show GPU info with CUDA version",
-        "",
-        "💡 After installing CUDA, please restart this setup.",
-        "",
-        "",
-    ],
-    'cuda_environment_issues': [
-        "",
-        "",
-        "📋 CUDA ENVIRONMENT CONFIGURATION:",
-        "",
-        "   CUDA toolkit appears to be installed, but environment configuration has issues.",
-        "   This prevents applications from finding and using CUDA properly.",
-        "",
-        "   Common environment problems:",
-        "      • CUDA compiler (nvcc) not in system PATH",
-        "      • CUDA_PATH environment variable not set",
-        "      • Multiple CUDA versions causing conflicts",
-        "      • Windows environment variables not updated after installation",
-        "      • Command prompt not restarted after CUDA installation",
-        "",
-        "   Why this matters:",
-        "      • Python packages like llama-cpp-python need to find CUDA during compilation",
-        "      • Without proper environment, GPU will be ignored even if CUDA is installed",
-        "",
-        "",
-    ],
-    'cuda_environment_fix': [
-        "",
-        "",
-        "📋 FIXING CUDA ENVIRONMENT CONFIGURATION:",
-        "",
-        "   Method 1 - Automatic Fix (Try First):",
-        "      1. Close all Command Prompt windows",
-        "      2. Restart your computer",
-        "      3. Open new Command Prompt",
-        "      4. Test commands:",
-        "            - nvcc --version",
-        "            - echo %CUDA_PATH%",
-        "      5. If this works, environment was just cached incorrectly",
-        "",
-        "   Method 2 - Manual Environment Variables:",
-        "      1. Find your CUDA installation directory:",
-        "         - Usually: C:\\Program Files\\NVIDIA GPU Computing Toolkit\\CUDA\\v12.x",
-        "         - Look for the newest version if multiple exist",
-        "",
-        "      2. Set environment variables:",
-        "         - Press Win+R, type 'sysdm.cpl', press Enter",
-        "         - Click 'Environment Variables'",
-        "         - In 'System Variables' section:",
-        "",
-        "      3. Add/Update CUDA_PATH:",
-        "         - Click 'New' (or select existing CUDA_PATH and click 'Edit')",
-        "         - Variable name: CUDA_PATH",
-        "         - Variable value: C:\\Program Files\\NVIDIA GPU Computing Toolkit\\CUDA\\v12.x",
-        "         - (Replace v12.x with your actual version)",
-        "",
-        "      4. Update PATH variable:",
-        "         - Find and select 'Path' in System Variables",
-        "         - Click 'Edit'",
-        "         - Add these paths (if not already present):",
-        "           * %CUDA_PATH%\\bin",
-        "           * %CUDA_PATH%\\libnvvp",
-        "         - Click 'OK' on all dialogs",
-        "",
-        "      5. Test the fix:",
-        "         - Close all Command Prompt windows",
-        "         - Open new Command Prompt",
-        "         - Run: nvcc --version",
-        "         - Run: echo %CUDA_PATH%",
-        "",
-        "   Method 3 - Reinstall CUDA (If Methods 1-2 Fail):",
-        "      1. Clean install:",
-        "         - Uninstall existing CUDA",
-        "         - Download latest CUDA toolkit",
-        "         - Restart computer when done",
-        "",
-    ],
-    'visual_studio_requirement_explanation': [
-        "",
-        "",
-        "📋 VISUAL STUDIO BUILD TOOLS REQUIREMENT:",
-        "",
-        "   This Monster Hunter game requires Visual Studio Build Tools to compile",
-        "   the llama-cpp-python package with CUDA support.",
-        "",
-        "   What needs to be installed:",
-        "      • Microsoft C++ compiler (MSVC)",
-        "      • Windows SDK for system headers",
-        "      • CMake and MSBuild tools",
-        "      • Total size: ~1-3 GB depending on options",
-        "",
-        "   SETUP:",
-        "      • The above checks may be false negative, ",
-        "      • If you believe you meet these requirements, choose [S] to skip this component",
-        "      • Otherwise, choose [G] to get further guidance",
-        "",
-        "💡 You only need 'Build Tools' - not the full Visual Studio IDE",
-        "",
-    ],
-    'visual_studio_installation_instructions': [
-        "",
-        "",
-        "📋 VISUAL STUDIO BUILD TOOLS INSTALLATION:",
-        "",
-        "   1. Go to https://visualstudio.microsoft.com/downloads/",
-        "",
-        "   2. Scroll down to 'All Downloads' section",
-        "",
-        "   3. Find and download 'Build Tools for Visual Studio 2022'",
-        "      (Look for the standalone 'Build Tools' - NOT the full IDE)",
-        "",
-        "   4. Run the downloaded installer (vs_buildtools.exe)",
-        "",
-        "   5. In the Visual Studio Installer, select workloads:",
-        "      •  Check 'C++ build tools' (this is the main workload you need)",
-        "",
-        "   6. In the 'Individual components' tab, ensure these are selected:",
-        "      •  MSVC v143 - VS 2022 C++ x64/x86 build tools (latest)",
-        "      •  Windows 11 SDK (10.0.22621.0) or Windows 10 SDK (latest)",
-        "      •  CMake tools for Visual Studio",
-        "",
-        "   7. Click 'Install' and wait for completion (10-30 minutes)",
-        "",
-        "   8. Restart your computer when installation finishes",
-        "",
-        "",
-    ],
-    'visual_studio_modify_for_cpp_tools': [
-        "",
-        "",
-        "📋 ADDING C++ TOOLS TO EXISTING VISUAL STUDIO:",
-        "",
-        "   You have Visual Studio installed but it's missing C++ compilation capabilities.",
-        "   Here's how to add them:",
-        "",
-        "   Method 1 - Visual Studio Installer (Recommended):",
-        "      1. Press Windows key and search for 'Visual Studio Installer'",
-        "      2. Launch the Visual Studio Installer",
-        "      3. Find your Visual Studio installation and click 'Modify'",
-        "      4. Go to the 'Workloads' tab",
-        "      5. Check the box for:",
-        "         • 'C++ build tools' (for Build Tools edition)",
-        "         OR",
-        "         • 'Desktop development with C++' (for Community/Professional edition)",
-        "",
-        "      6. Click 'Individual components' tab and verify these are selected:",
-        "         • MSVC v143 - VS 2022 C++ x64/x86 build tools",
-        "         • Windows 11 SDK or Windows 10 SDK (latest version)",
-        "         • CMake tools for Visual Studio",
-        "",
-        "      7. Click 'Modify' and wait for installation (5-20 minutes)",
-        "",
-        "      8. Restart your computer after installation completes",
-        "",
-        "   Method 2 - Fresh Install (If Modify Doesn't Work):",
-        "      1. Uninstall current Visual Studio from Windows Settings",
-        "      2. Download Build Tools for Visual Studio 2022",
-        "      3. Follow the fresh installation instructions",
-        "",
-        "💡 If you can't find Visual Studio Installer:",
-        "   - Download it again from https://visualstudio.microsoft.com/downloads/",
-        "   - Look for 'Visual Studio Installer' in the Tools section",
-        "",
-    ],
-    'visual_studio_missing_cpp_tools_detected': [
-        "",
-        "📋 VISUAL STUDIO MISSING C++ BUILD TOOLS:",
-        "",
-        "   Visual Studio is installed on your system, but it doesn't include C++ compilation capabilities.",
-        "   This is a common configuration when Visual Studio was installed for other development",
-        "   (like web development, .NET, or Python) without including C++ tools.",
-        "",
-        "   What you have:",
-        "      ✅ Visual Studio installation",
-        "      ✅ Windows SDK",
-        "      ❌ C++ build tools (MSVC compiler)",
-        "",
-        "   What you need:",
-        "      • Add the C++ build tools workload to your existing Visual Studio installation",
-        "",
-        "💡 This can be fixed by modifying your current Visual Studio installation",
-        "",
-    ],
-    'visual_studio_sdk_detection_issue': [
-        "",
-        "📋 WINDOWS SDK DETECTION ISSUE:",
-        "",
-        "   Visual Studio and C++ build tools are detected, but Windows SDK was not found",
-        "   in the expected locations. This is likely a false negative.",
-        "",
-        "   What you have:",
-        "      ✅ Visual Studio installation",
-        "      ✅ C++ build tools (MSVC compiler)",
-        "      ❌ Windows SDK (may be bundled with Visual Studio)",
-        "",
-        "   Most likely cause:",
-        "      • Windows SDK is included with your Visual Studio installation",
-        "      • SDK components are in VS directory instead of standalone location",
-        "",
-        "💡 Your build environment is probably functional despite this detection issue",
-        "",
-    ],
-    'visual_studio_partial_installation_detected': [
-        "",
-        "📋 PARTIAL DEVELOPMENT ENVIRONMENT DETECTED:",
-        "",
-        "   Some development components were found but Visual Studio is not installed",
-        "   in the expected locations. This suggests an incomplete or non-standard setup.",
-        "",
-        "   What was detected:",
-        "      ❌ Visual Studio installation (not found in standard directories)",
-        "      ✅ Some development components (Windows SDK or build tools)",
-        "",
-        "   Possible causes:",
-        "      • Visual Studio installed in non-standard location",
-        "      • Standalone Windows SDK installation without Visual Studio",
-        "      • Incomplete or corrupted Visual Studio installation",
-        "",
-        "💡 A complete Visual Studio Build Tools installation is recommended",
-        "",
-    ],
-    'llama_cpp_requirement_explanation': [
-        "",
-        "📋 LLM INTEGRATION REQUIREMENT:",
-        "",
-        "   This Monster Hunter game requires llama-cpp-python for AI language model inference.",
-        "   The game's monsters, NPCs, and story generation depend on fast AI responses.",
-        "",
-        "   Performance requirements:",
-        "      ⚡ Game designed for: ~35+ tokens/second (smooth gameplay)",
-        "      🐌 CPU-only typically: <3 tokens/second (game becomes unplayable)",
-        "",
-        "   What you're choosing:",
-        "      • CUDA version: Compiles with your NVIDIA GPU for fast inference",
-        "      • CPU version: Uses only your processor (extremely slow for gaming)",
-        "",
-        "   Installation complexity:",
-        "      • CUDA installation tries 3 different methods automatically",
-        "      • May take 5-15 minutes depending on your system",
-        "      • Requires Visual Studio Build Tools and CUDA toolkit",
-        "",
-        "💡 CUDA acceleration is essential for playable game performance",
-        "",
-    ],
-    'llama_cpp_cuda_install_failed': [
-        "",
-        "📋 CUDA INSTALLATION FAILED:",
-        "",
-        "   The automatic CUDA installation encountered problems.",
-        "   This usually indicates missing prerequisites or system configuration issues.",
-        "",
-        "   Common causes:",
-        "      • Visual Studio Build Tools not properly installed",
-        "      • CUDA toolkit missing or not in PATH",
-        "      • NVIDIA drivers too old or missing",
-        "      • Network issues downloading packages",
-        "      • Insufficient disk space for compilation",
-        "",
-        "   What you can do:",
-        "      1. Retry installation (sometimes temporary issues resolve)",
-        "      2. Install CPU-only version (playable but very slow)",
-        "      3. Skip for now and troubleshoot prerequisites",
-        "",
-        "💡 CPU-only version lets you test the game while fixing CUDA issues",
-        "",
-    ],
-    'llama_cpp_cpu_warning': [
-        "",
-        "⚠️  CPU-ONLY INSTALLATION WARNING:",
-        "",
-        "   You're about to install the CPU-only version of llama-cpp-python.",
-        "   This will make the game extremely slow and frustrating to play.",
-        "",
-        "   Expected performance impact:",
-        "      • Token generation: Typically <3 tokens/second (vs 35+ needed)",
-        "      • Game timeouts: Prompts will frequently timeout and retry",
-        "      • Gameplay flow: Constant interruptions from slow generation",
-        "      • Overall experience: Nearly unplayable due to performance issues",
-        "",
-        "   Technical issues with slow generation:",
-        "      • Game assumes broken AI when prompts take too long",
-        "      • Automatic retries can create endless loops",
-        "      • Even with adjusted timeout settings, gameplay remains unbearably slow",
-        "",
-        "   Why continue with CPU-only:",
-        "      ✓ Test that the game installation works",
-        "      ✓ Verify other components before fixing GPU acceleration",
-        "      ✓ Temporary solution while troubleshooting CUDA",
-        "",
-        "   You can upgrade to CUDA later by running setup again.",
-        "",
-        "💡 This is only recommended for testing, not actual gameplay",
-        "",
-    ],
-    'llama_cpp_broken_installation': [
-        "",
-        "📋 BROKEN INSTALLATION DETECTED:",
-        "",
-        "   llama-cpp-python is installed but cannot be imported properly.",
-        "   This usually indicates a corrupted or incomplete installation.",
-        "",
-        "   Common causes:",
-        "      • Installation was interrupted",
-        "      • Mixed CUDA and CPU versions installed",
-        "      • Missing dependencies or DLLs",
-        "      • Virtual environment corruption",
-        "      • Python version compatibility issues",
-        "",
-        "   Solution:",
-        "      • Clean reinstallation is the most reliable fix",
-        "      • This will remove the broken installation and start fresh",
-        "",
-        "💡 Reinstallation usually resolves import and compatibility issues",
-        "",
-    ],
-    'llama_cpp_cpu_detected': [
-        "",
-        "📋 CPU-ONLY PERFORMANCE DETECTED:",
-        "",
-        "   Performance test indicates llama-cpp-python is running on CPU only.",
-        "   This will make the game extremely slow and potentially broken.",
-        "",
-        "   Current performance:",
-        "      • Very slow token generation (under 3 tokens/second)",
-        "      • Game designed for 35+ tokens/second performance",
-        "      • Prompts will frequently timeout and retry",
-        "      • Game may become stuck in retry loops",
-        "",
-        "   Technical issues:",
-        "      • Game assumes AI is broken when generation is too slow",
-        "      • Timeout settings can be adjusted but don't solve core speed issue",
-        "      • Even with maximum timeouts, gameplay becomes unbearably slow",
-        "",
-        "   Why this happened:",
-        "      • CUDA installation failed silently",
-        "      • GPU drivers not working properly",
-        "      • CUDA libraries not found at runtime",
-        "      • CPU-only version was installed intentionally",
-        "",
-        "   Recommendations:",
-        "      • Upgrade to CUDA version for playable performance",
-        "      • Keep CPU version only for testing purposes",
-        "",
-        "💡 The game is designed for GPU acceleration and needs it to be fun",
-        "",
-    ],
-    'llama_cpp_weak_gpu_detected': [
-        "",
-        "📋 WEAK GPU PERFORMANCE DETECTED:",
-        "",
-        "   Performance test shows slower than expected GPU acceleration.",
-        "   The game should still be playable but may feel sluggish at times.",
-        "",
-        "   Current performance:",
-        "      • Moderate token generation (3-10 tokens/second)",
-        "      • Below optimal 35+ tokens/second target",
-        "      • Prompts will frequently timeout and retry",
-        "      • Game may become stuck in retry loops",
-        "",
-        "   Possible causes:",
-        "      • Older or entry-level GPU",
-        "      • Not all model layers loaded on GPU",
-        "      • GPU memory limitations",
-        "      • Suboptimal CUDA installation",
-        "",
-        "   Options:",
-        "      • Try CUDA reinstall (may improve GPU utilization)",
-        "      • Install a smaller faster model (quality will drop)",
-        "      • Adjust model settings later for better performance",
-        "",
-        "💡 You may need to adjust timeout limits in lmm config",
-        "",
-    ],
-    'llm_model_requirement_explanation': [
-        "",
-        "📋 LLM MODEL CONFIGURATION REQUIRED:",
-        "",
-        "   This Monster Hunter game requires a language model file for AI-powered",
-        "   monster interactions, NPC conversations, and dynamic story generation.",
-        "",
-        "   What you need:",
-        "      • A GGUF format language model (recommended: 7B or 13B parameters)",
-        "      • Model file stored locally on your computer",
-        "      • At least 8GB free disk space for the model",
-        "      • Model configured in the .env file for the game to find it",
-        "",
-        "   Recommended Model: Kunoichi 7B",
-        "",
-        "   Other popular model options:",
-        "      • Llama 2 7B Chat (good balance of speed and quality)",
-        "      • Mistral 7B Instruct (fast and capable)",
-        "      • CodeLlama 7B Instruct (good for technical content)",
-        "",
-        "   Where to get models:",
-        "      • Hugging Face (search for GGUF format)",
-        "      • TheBloke's model collection (pre-quantized GGUF)",
-        "      • Local model conversion tools",
-        "",
-        "💡 The game is designed to work with most 7B+ parameter models in GGUF format",
-        "",
-    ],
-    'llm_model_path_placeholder': [
-        "",
-        "📋 SETTING LLM MODEL PATH:",
-        "",
-        "   Please enter the full path to your model file",
-        "",
-        "   Example valid paths:",
-        "      C:/Users/YourName/.cache/lm-studio/models/TheBloke/Kunoichi-7B-GGUF/kunoichi-7b.Q6_K.gguf",
-        "      D:/Downloads/mistral-7b-instruct-v0.1.Q4_K_M.gguf",
-        "      C:/AI/models/llama-2-7b-chat.q4_K_M.gguf",
-        "",
-        "💡 Use forward slashes (/) in the path, even on Windows",
-        "",
-    ],
-    'llm_model_file_missing': [
-        "",
-        "📋 MODEL FILE NOT FOUND:",
-        "",
-        "   Your .env file has a model path configured, but the file doesn't exist",
-        "   at that location. The file may have been moved, deleted, or the path",
-        "   may contain a typo.",
-        "",
-        "   Common solutions:",
-        "      • Check if the file was moved to a different folder",
-        "      • Verify the filename is spelled correctly (including extension)",
-        "      • Make sure the drive letter is correct (C:/, D:/, etc.)",
-        "      • Check if the file was accidentally deleted",
-        "",
-        "   Model file requirements:",
-        "      • Must be in GGUF format (.gguf extension)",
-        "      • Must be at least 100MB in size",
-        "      • Must be accessible/readable by the game",
-        "",
-        "💡 You can download models from Hugging Face or TheBloke's collection",
-        "",
-    ],
-    'llm_model_path_invalid': [
-        "",
-        "📋 INVALID MODEL PATH CONFIGURATION:",
-        "",
-        "   There's an issue with your current model path configuration.",
-        "   This could be due to various reasons like incorrect file type,",
-        "   permission issues, or file corruption.",
-        "",
-        "   Common issues:",
-        "      • File is not a valid model format (.gguf, .ggml, .bin, .safetensors)",
-        "      • File is too small to be a language model (< 100MB)",
-        "      • Permission denied when trying to access the file",
-        "      • File appears to be corrupted or incomplete",
-        "",
-        "   What to check:",
-        "      • File extension should be .gguf for best compatibility",
-        "      • File size should be several GB for good models",
-        "      • File should be in a location you have read access to",
-        "",
-        "💡 GGUF format models from Hugging Face are most reliable",
-        "",
-    ],
-    'llm_model_download_guidance': [
-        "",
-        "📋 HOW TO FIND AND DOWNLOAD A LANGUAGE MODEL:",
-        "",
-        "   Method 1 - Hugging Face (Direct Download):",
-        "      • Go to https://huggingface.co/",
-        "      • Search for models by 'TheBloke' (pre-quantized GGUF models)",
-        "      • Good starting models: 'Llama-2-7B-Chat-GGUF', 'Mistral-7B-Instruct-v0.1-GGUF'",
-        "      • Click model → 'Files and versions' tab → download the .gguf file",
-        "      • Choose Q4_K_M version for best balance of quality and speed",
-        "",
-        "   Method 2 - LM Studio (Easiest for Beginners):",
-        "      • Download LM Studio from https://lmstudio.ai/",
-        "      • Install and open LM Studio",
-        "      • Go to 'Discover' tab and search for models",
-        "      • Download any model you like (LM Studio handles GGUF automatically)",
-        "      • Models are saved to: C:/Users/YourName/.cache/lm-studio/models/",
-        "",
-        "   Understanding Quantization (Q4_K_M, Q5_K_M, etc.):",
-        "      • Quantization = compression technique to reduce model size",
-        "      • Q4_K_M: ~4GB, good quality, fast (recommended for gaming)",
-        "      • Q5_K_M: ~5GB, better quality, slower",
-        "      • Q8_0: ~7GB, highest quality, slowest",
-        "      • Start with Q4_K_M unless you have plenty of VRAM",
-        "",
-        "   File Management:",
-        "      • Save models to a dedicated folder like C:/AI/models/ or D:/models/",
-        "      • Models are typically 3-8GB each, so ensure you have disk space",
-        "      • Use descriptive folder names if you plan to try multiple models",
-        "      • The exact folder location doesn't matter as long as you remember it",
-        "",
-        "💡 LM Studio is the easiest way to get started - it handles everything automatically",
-        "",
-    ],
-    'basic_backend_diagnostic': [
-        "",
-        "🚨 CRITICAL: BASIC BACKEND SETUP FAILED 🚨",
-        "",
-        "⚠️  THE GAME CANNOT BE SETUP WITH THESE ISSUES ⚠️",
-        "",
-        "   Basic backend components are REQUIRED PREREQUISITES for all other setup steps.",
-        "   If you continue with these issues, ALL other components will fail:",
-        "",
-        "   Common problems and solutions:",
-        "",
-        "   Python Version Issues:",
-        "      • Python 3.8+ required - upgrade from https://python.org/downloads",
-        "      • Make sure to check 'Add Python to PATH' during installation",
-        "      • Restart command prompt after Python installation",
-        "",
-        "   pip Package Manager Issues:",
-        "      • pip should come with Python - if missing, Python installation is broken",
-        "      • Try: python -m ensurepip --upgrade",
-        "      • Or reinstall Python with default settings",
-        "",
-        "   Network Access Issues:",
-        "      • Check internet connection for downloading packages",
-        "",
-        "   Virtual Environment Issues:",
-        "      • Usually auto-created, failure indicates permission or disk space problems",
-        "      • Check you have write permissions in current directory",
-        "      • Antivirus may be blocking file creation",
-        "",
-        "   Environment File Issues:",
-        "      • .env.example template file may be missing from project",
-        "",
-        "   Why continuing is NOT recommended:",
-        "      • Other components depend on working Python environment",
-        "      • Installation will likely fail at later stages",
-        "      • May cause confusing error messages",
-        "      • Game won't work without proper backend setup",
-        "",
-        "   RECOMMENDED ACTION:",
-        "      1. Exit this setup now",
-        "      2. Fix the issues shown above",
-        "      3. Restart your command prompt",
-        "      4. Run this setup again",
-        "      5. The auto-setup should work once environment is fixed",
-        "",
-        "⚠️  ONLY CONTINUE IF YOU ARE AN EXPERIENCED DEVELOPER ⚠️",
-        "",
-    ],
-}
-
-
-def get_message(key):
-    """
-    Get user messages by key
-
-    Args:
-        key (str): Message key from MESSAGES dict
-
-    Returns:
-        list: List of message lines, or empty list if key not found
-    """
-    return MESSAGES.get(key, [])
-
-
-def get_available_messages():
-    """
-    Get list of all available messages keys
-
-    Returns:
-        list: All message keys
-    """
-    return list(MESSAGES.keys())
